+++ pkg/querier/tripperware/queryrange/query_range_middlewares_test.go
@@ -34,6 +34,8 @@ func TestRoundTrip(t *testing.T) {
 					_, err = w.Write([]byte(responseBody))
 				} else if r.RequestURI == queryWithWarnings {
 					_, err = w.Write([]byte(responseBodyWithWarnings))
+				} else if r.RequestURI == queryWithInfos {
+					_, err = w.Write([]byte(responseBodyWithInfos))
 				} else {
 					_, err = w.Write([]byte("bar"))
 				}